\name{qrjoint-internal}
\Rdversion{0.1-1}
\alias{estFn}
\alias{ppFn0}
\alias{ppFn}
\alias{lamFn}
\alias{nuFn}
\alias{nuFn.inv}
\alias{sigFn}
\alias{sigFn.inv}
\alias{unitFn}
\alias{q0}
\alias{Q0}
\alias{F0}
\alias{sum.sq}
\alias{extract}
\alias{logmean}
\alias{logsum}
\alias{shrinkFn}
\alias{trape}
\alias{klGP}
\alias{proxFn}

\title{Internal qrjoint Functions}
\description{These functions are repeatedly used by the main functions of the package "qrjoint". Not intended for direct use the by user.} 

\keyword{internal}